pyproject.toml
[tool.poetry]
name = "opendataqna"
version = "0.1.0"
description = ""
authors = ["msubasioglu","l-sri","orpzs"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.9.7 || >3.9.7, <4.0"
google = "^3.0.0"
google-api-python-client = "^2.120.0"
google-cloud = "^0.34.0"
google-cloud-aiplatform = "^1.43.0"
pandas = "^2.2.1"
pandas-gbq = "^0.21.0"
cloud-sql-python-connector = {extras = ["asyncpg"], version = "^1.7.0"}
sqlalchemy = "^2.0.27"
pgvector = "^0.2.5"
pg8000 = "^1.30.5"
psycopg2-binary = "^2.9.9"
tabulate = "^0.9.0"
langchain = "^0.1.9"
fsspec = "^2024.2.0"
gcsfs = "^2024.2.0"
streamlit = "^1.32.2"
google-cloud-bigquery = "^3.17.2"
google-cloud-bigquery-connection = "^1.15.3"
pyarrow = "^15.0.2"
Flask = { version = "^2.1.0", extras = ["async"]}
gunicorn = "^20.1.0"
flask-cors = "^4.0.0"
Werkzeug = "^2.3.7"
google-cloud-firestore= "^2.16.0"
firebase-admin = "^5.2.0"
sentence-transformers = "^2.6.1"



[tool.poetry.group.dev.dependencies]
pyzmq ="<26.0.0"
ipykernel = "^6.29.3"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"